+++ internal/access/manager/conversations.go
@@ -47,11 +47,11 @@ type RecentConversationDTO struct {
 	// Timestamp is the latest message timestamp in Unix seconds.
 	Timestamp int64 `json:"timestamp"`
 	// LastMsgSeq is the latest message sequence known to conversation sync.
-	LastMsgSeq uint32 `json:"last_msg_seq"`
+	LastMsgSeq uint64 `json:"last_msg_seq"`
 	// LastClientMsgNo is the latest client message number when present.
 	LastClientMsgNo string `json:"last_client_msg_no"`
 	// ReadToMsgSeq is UID's read cursor for this conversation.
-	ReadToMsgSeq uint32 `json:"read_to_msg_seq"`
+	ReadToMsgSeq uint64 `json:"read_to_msg_seq"`
 	// Version is the sync compatibility version timestamp.
 	Version int64 `json:"version"`
 	// RecentMessages contains newest message previews for this conversation.
